requirements.txt
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
curl_cffi==0.10.0
cycler==0.12.1
fonttools==4.58.0
frozendict==2.4.6
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
multitasking==0.0.11
mysql-connector-python==9.3.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
protobuf==5.29.4
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
seaborn==0.13.2
six==1.17.0
soupsieve==2.7
ta==0.11.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
yfinance==0.2.59
